Replace u64 -> uint, i64 -> int

+++ 0001-0050/prob0009.rs
@@ -1,10 +1,10 @@
 extern mod euler;
 use arith = euler::arith;
 
-fn find_pyrhagorean(sum: u64) -> ~[(u64, u64, u64)] {
+fn find_pyrhagorean(sum: uint) -> ~[(uint, uint, uint)] {
     let mut answer = ~[];
-    for u64::range(2u64, sum - 2u64) |c| {
-        for u64::range(1u64, u64::min((sum - c) / 2u64, arith::isqrt(c*c / 2u64))) |a| {
+    for uint::range(2, sum - 2) |c| {
+        for uint::range(1, uint::min((sum - c) / 2, arith::isqrt(c*c / 2))) |a| {
             let b = sum - c - a;
             if a * a + b * b == c * c {
                 answer += [(a, b, c)];
@@ -15,9 +15,9 @@ fn find_pyrhagorean(sum: u64) -> ~[(u64, u64, u64)] {
 }
 
 fn main() {
-    for find_pyrhagorean(1000u64).each() |tp| {
+    for find_pyrhagorean(1000).each() |tp| {
         let (a, b, c) = *tp;
-        io::println(#fmt("%u^2 + %u^2 = %u^2", a as uint, b as uint, c as uint));
-        io::println(#fmt("prod: %u", a * b * c as uint));
+        io::println(#fmt("%u^2 + %u^2 = %u^2", a, b, c));
+        io::println(#fmt("prod: %u", a * b * c));
     }
 }
