+++ 0001-0050/prob0010.rs
@@ -3,12 +3,12 @@ extern mod euler;
 use prime = euler::prime;
 
 fn main() {
-    let mut sum = 0u64;
+    let mut sum = 0;
     for prime::Prime().each |p| {
-        if p >= 2000000u64 {
+        if p >= 2000000 {
             break;
         }
         sum += p;
     }
-    io::println(#fmt("%u", sum as uint));
+    io::println(#fmt("%u", sum));
 }
